Fix a postgresql warning about OID 705 when running rspec tests
This warning is diplayed when running test with rspec and postgresql :
unknown OID 705: failed to recognize type of 'read_mark_reader_type'. It will be treated as String.

Reading this rails PR : https://github.com/rails/rails/pull/10498/files the
fix for postgresql is to cast values when using 'AS' in custom queries